src/requirements.txt
Flask_Login==0.4.1
SQLAlchemy==1.2.9
Flask_SQLAlchemy==2.3.2
Flask_WTF==0.14.2
Flask_Bcrypt==0.7.1
WTForms==2.2.1
alembic==1.0.0
Flask_Migrate==2.2.1
Flask_Mail==0.9.1
Flask_Script==2.0.6
Flask==1.0.2
itsdangerous==0.24
Pillow==5.2.0
unicode_slugify==0.1.3
pymysql
